refactor(runtime): four pure-move splits out of the largest files
Nothing here changed but the file it lives in. Each block was cut
whole, with its doc comments, and pasted into a child module of the
file it came from, so private fields and imports stay reachable and
no signature or visibility beyond `pub(super)` had to move:

- `fanout.rs` (3,527 lines): the results report and the status helpers
  (`build_report`, `inject_results`, `fit_worker_section`,
  `bytes_per_worker`, `agent_status`, `set_status` and their three
  constants) are `fanout/report.rs`.
- `pipeline/transition.rs`: `gate_blocks` and `spend_gate_attempt` are
  `pipeline/gate_check.rs`, re-exported at crate visibility from
  `pipeline` as `transition`'s items already were.
- `world.rs` (3,001 lines): the quiescence helpers (`count`,
  `agent_digest`, `fingerprint`, `has_async_inflight`) are
  `world/quiescence.rs`.
- `components/context_window.rs` (1,698 lines): `try_evict`, the
  eviction cascade, is `components/context_window/eviction.rs`.

Verified as pure moves: for each file, the multiset of non-comment,
non-import lines it lost equals the multiset the new file carries,
after stripping the `pub(super)` the moved items needed. The runtime's
1,713 tests run unchanged.

(cherry picked from commit 2e64315d426eb72723048b20a860a8a511ea4598)
