+++ crates/leviath-runtime/src/components/context_window.rs
@@ -12,6 +12,8 @@ use super::block_cache::{
 };
 use super::*;
 
+mod eviction;
+
 /// Result of an eviction attempt, including tokens freed and regions needing LLM compaction.
 #[derive(Debug, Clone)]
 pub struct EvictionResult {
@@ -382,210 +384,6 @@ impl ContextWindow {
     pub fn needs_eviction(&self, threshold: f32) -> bool {
         let usage_ratio = self.current_tokens as f32 / self.max_tokens as f32;
         usage_ratio >= threshold
-    }
-
-    /// Execute eviction cascade to free up space.
-    ///
-    /// Returns an `EvictionResult` with tokens freed and any regions that need
-    /// LLM-based compaction. The caller is responsible for performing compaction
-    /// on the listed regions (since it requires async LLM access).
-    pub fn try_evict(&mut self, target_free_tokens: usize) -> leviath_core::Result<EvictionResult> {
-        use leviath_core::RegionKind;
-
-        let initial_tokens = self.current_tokens;
-
-        // A region under `admission = "reject"` is exempt from every phase
-        // below. Refusing writes to protect what a region holds would mean
-        // nothing if the window-level cascade could take the same entries a
-        // moment later - `reject` would only change which code did the silent
-        // dropping. The agent releases from these, or nothing does.
-        let evictable = |r: &Region| {
-            r.admission != leviath_core::region::Admission::Reject
-                && matches!(
-                    r.kind,
-                    RegionKind::Clearable
-                        | RegionKind::Temporary
-                        | RegionKind::Custom {
-                            persistent: false,
-                            ..
-                        }
-                )
-        };
-
-        // Check if we have any evictable regions
-        let has_evictable = self.regions.iter().any(evictable);
-
-        if !has_evictable {
-            tracing::warn!(
-                "Context window has no Clearable or Temporary regions. \
-                 This may be intentional, but usually indicates a configuration error."
-            );
-        }
-
-        // Phase 1: Clear Clearable regions (all-or-nothing)
-        for region in &mut self.regions {
-            if matches!(region.kind, RegionKind::Clearable)
-                && region.admission != leviath_core::region::Admission::Reject
-                && !region.content.is_empty()
-            {
-                let freed = region.current_tokens;
-                region.clear();
-                self.current_tokens -= freed;
-                tracing::debug!(
-                    region = %region.name,
-                    tokens_freed = freed,
-                    "Cleared Clearable region (all-or-nothing)"
-                );
-
-                if self.max_tokens.saturating_sub(self.current_tokens) >= target_free_tokens {
-                    return Ok(EvictionResult {
-                        tokens_freed: initial_tokens - self.current_tokens,
-                        needs_compaction: Vec::new(),
-                    });
-                }
-            }
-        }
-
-        // Phase 1.5: Give each non-persistent custom region's on_overflow
-        // hook first say over what IT loses, before the indiscriminate
-        // oldest-first cascade below. A script that keeps errors and drops
-        // successes only works if it runs before oldest-first does. Hook
-        // absent/failing/insufficient → phase 2 makes the guaranteed
-        // progress.
-        let mut custom_freed = 0usize;
-        for i in 0..self.regions.len() {
-            let needed = target_free_tokens
-                .saturating_sub(self.max_tokens.saturating_sub(self.current_tokens));
-            if needed == 0 {
-                break;
-            }
-            let region = &self.regions[i];
-            if !matches!(
-                region.kind,
-                RegionKind::Custom {
-                    persistent: false,
-                    ..
-                }
-            ) || region.admission == leviath_core::region::Admission::Reject
-                || region.content.is_empty()
-            {
-                continue;
-            }
-            let Some(script) = self.custom_script_for(&region.name.clone()) else {
-                continue;
-            };
-            if !script.has_on_overflow() {
-                continue;
-            }
-            let freed = crate::custom_region::apply_overflow(&script, &mut self.regions[i], needed);
-            self.current_tokens = self.current_tokens.saturating_sub(freed);
-            custom_freed += freed;
-            if freed > 0 {
-                tracing::debug!(
-                    region = %self.regions[i].name,
-                    tokens_freed = freed,
-                    "custom region's on_overflow chose its own evictions"
-                );
-            }
-        }
-        // Return early ONLY when a script's own drops satisfied the target -
-        // otherwise phase 2 would immediately evict one more entry (it checks
-        // the target *after* each eviction), overriding the script's
-        // retention choice. Windows with no custom drops (custom_freed == 0)
-        // fall through with phase 2's pre-existing behavior, byte-identical.
-        if custom_freed > 0
-            && self.max_tokens.saturating_sub(self.current_tokens) >= target_free_tokens
-        {
-            return Ok(EvictionResult {
-                tokens_freed: initial_tokens - self.current_tokens,
-                needs_compaction: Vec::new(),
-            });
-        }
-
-        // Phase 2: Evict from Temporary regions (oldest first, one at a time).
-        // Non-persistent Custom regions join this phase: their script's
-        // on_overflow hook (when present) has already had its say in phase
-        // 1.5; oldest-first is the guaranteed-progress fallback.
-        loop {
-            let mut evicted_any = false;
-
-            for region in &mut self.regions {
-                if matches!(
-                    region.kind,
-                    RegionKind::Temporary
-                        | RegionKind::Custom {
-                            persistent: false,
-                            ..
-                        }
-                ) && region.admission != leviath_core::region::Admission::Reject
-                    && let Some(entry) = region.remove_oldest()
-                {
-                    let freed = entry.tokens;
-                    self.current_tokens -= freed;
-                    evicted_any = true;
-
-                    tracing::debug!(
-                        region = %region.name,
-                        tokens_freed = freed,
-                        "Evicted temporary region entry (oldest first)"
-                    );
-
-                    if self.max_tokens.saturating_sub(self.current_tokens) >= target_free_tokens {
-                        return Ok(EvictionResult {
-                            tokens_freed: initial_tokens - self.current_tokens,
-                            needs_compaction: Vec::new(),
-                        });
-                    }
-                }
-            }
-
-            if !evicted_any {
-                break;
-            }
-        }
-
-        // Phase 3: If still need space, identify Compacting regions that need compaction
-        let mut needs_compaction = Vec::new();
-        if self.max_tokens.saturating_sub(self.current_tokens) < target_free_tokens {
-            for region in &self.regions {
-                if region.needs_compaction() {
-                    needs_compaction.push(region.name.clone());
-                }
-            }
-        }
-
-        // Phase 4: SlidingWindow regions are NEVER reduced
-        // Phase 5: Pinned and CompactHistory regions are NEVER touched
-
-        // Check for pinned regions over budget
-        let pinned_tokens: usize = self
-            .regions
-            .iter()
-            .filter(|r| {
-                matches!(
-                    r.kind,
-                    RegionKind::Pinned
-                        | RegionKind::CompactHistory { .. }
-                        | RegionKind::Custom {
-                            persistent: true,
-                            ..
-                        }
-                )
-            })
-            .map(|r| r.current_tokens)
-            .sum();
-
-        if pinned_tokens > self.max_tokens {
-            return Err(leviath_core::Error::PinnedRegionsOverBudget {
-                pinned_tokens,
-                total_budget: self.max_tokens,
-            });
-        }
-
-        Ok(EvictionResult {
-            tokens_freed: initial_tokens - self.current_tokens,
-            needs_compaction,
-        })
     }
 
     /// Result of assembling the context window into system blocks + messages.
